microservices/classification_service/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
python-multipart==0.0.5
fireo==1.4.1
google-cloud-bigquery==2.20.0
google-cloud-storage==1.42.2

appdirs==1.4.4
backports-datetime-fromisoformat==1.0.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
Faker==8.5.1
firebase-admin==4.5.3
fireo==1.4.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.2
google-cloud-firestore==1.9.0
google-cloud-secret-manager==2.5.0
google-cloud-aiplatform==1.9.0
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.9.0
libcst==0.3.19
msgpack==1.0.2
mypy-extensions==0.4.3
neo4j==4.3.1
numpy==1.21.1
packaging==21.0
pandas==1.2.4
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.2.2
pyparsing==2.4.7
pyppeteer==0.2.6
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tqdm==4.62.3
typing-extensions==3.10.0.0
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.6
websockets==9.1
zipp==3.6.0

# pdf splitter
pypdf2==1.26.0

# pdf 2 image
pdf2image==1.15.0

# computer vision
opencv-python==4.1.2.30
PyMuPDF